Adds factory method with validation to GlobalViewParameters
In order to make sure the used locale is available, URLs are configured
for each locale and URLs do not have unconfigured URLs, validation is applied.
This occurs in a separate method, because configuration validation should not
happen in the constructor and because we do not want to imply locales are
stored internally.