Adds specified consent and specified consent list
Also updates/renames the service that uses it, SpecifiedConsentListService,
updates the controller that consumes this service and the layout
that renders them.

The service no longer requires knowledge over the userProvider,
as a specified consent list can be retrieved for a given user.

A SpecifiedConsent contains Consent and the Attribute for which
this Consent is given.
In the future, we might work with an API to retrieve which attributes
are specifically released to each SP. Then, the concepts of
SpecifiedConsent and Consent could be merged.