Show ORCID in the template
The AttributeAggregationService was added who is tasked with finding
the enabled AAattributes for the currently logged in user. To search
the attribute, AA Api finds by SurfConextUserId, this user id is
loaded from the attributes of the currently logged in user.

Preparations have been taken to return a list of AAAttributes, this
list is not implemented yet.