Load attribute sources alongside their values
This commit makes sure the attribute sources are also considered when
showing the my services overview. The optional 'showSources' parameter
is posted to the arp endpoint, making it return the sources of the
attributes that are fed to the ARP.

The changed response was reflected in the SpecifiedConsent value object.

Changes in the overview template include the distinction between idp
and aa attribute overviews.

Disclaimer: did a little fixup work in this commit due to some rebase
issues.