Filter blacklisted attributes
The urn:oid:1.3.6.1.4.1.1076.20.40.40.1 and
urn:oid:1.3.6.1.4.1.1466.115.121.1.15 attributes are not to be shown to
end users. These attributes are only kept for historic reasons. So
filter them when retrieving their attributes. Leaving them in the set
of stored attributes. This is required to load the consent data for the
logged in user. As consent is loaded from engine block by
urn:oid:1.3.6.1.4.1.1076.20.40.40.1 (surfconextid).

A test was added to ascertain the attribute filtering actually works.