Support account list in attribute aggregator response
Asking the attribute aggregator for a list of connected accounts
returns a response like:

    [
       {
          "accountType" : "ORCID",
          "created" : 1509708987,
          "id" : 50,
          "linkedId" : "http://orcid.org/0000-0002-5385-9420",
          "schacHome" : "surfnet.nl",
          "urn" : "urn:collab:person:surfnet.nl:kinkhorst"
       }
    ]

Previous implementation expected an ARP reply, which is not how
profile obtains the connection status of linked accounts. Major
difference is that we now configure the accountType (ORCID) instead of
the urn of the attribute and match that to the account list.