Read Arp configuration to show AA attributes
The arp configuration is read from the EB API. The response transformed
to an Arp value object and is loaded on the SpecifiedConsent value
object. The SpecifiedConsent value object is used in the view to load
the categorized (my attribute source) attribute information.

Note that the aggregated attributes are not shown with a value. This
value can/will be added at a later stage in the project.