Implement active/available connections
This entailed some changes to the way the active and available
connections are displayed and handled by the controller. The new
solution is a step up from the previous solution in terms of support for
future connections.

In the previous version of the overview.html.twig, the template checked
if the orcid connection was connected or not. This logic is moved to the
AttributeAggregationAttributesList object. It can be used to fetch a
list of active and available attributes (connections). Relieving the
template of this logic.

Two new translations where added as requested in the story.