Utilize read method for retrieving ARP config
Previously the ARP config was retrieved using a POST request. This was
copied from the /arp POST call. The /arp request posts a potentially
large JSON blob in its request, making it unsuitable for a GET request.

This is not the case for the read-arp request. As it only requires a
list of entity ids to be passed.

So a comma seperated list of entityIds is passed using a get parameter.

The url encoding of the parameter is handled by Guzzle.