Create the ConfirmConnectionDeleteType form
A simple form was created consisting of a checkbox and a sumbit button.
Some styling and javascript was applied to the form to enhance the user
experience.

Translations have been added for the form labels.

For now this form can used for just one connection. Some refactoring
might be required when other connections are to be used.