Allow IDP metadata without contacts
The ContactPersonList initialized in an invalid state when created
with an empty array. It is now possible to have a list object without
contacts in it.

The validation on the contact list sent by EngineBlock was invalid and
has been restored in this commit.

Fixes #94
See: https://www.pivotaltracker.com/story/show/155390093
And: https://www.pivotaltracker.com/story/show/111038212