Stop using SURFconextId as user identifier
The nameId value is now used instead. Containing the exact same value
as long as the NameID was configured to be unspecified (which it is in
our case).

The NameID can already be read from the AuthenticatedUser object, so
manually retrieving it from the assertion subject is not required.

The ConsentService could be cleaned up drastically.

https://www.pivotaltracker.com/story/show/164103691