Remove unrealistic test cases
First, these tests are acually testing the StepUp SAML bundle and not
really any features of Profile. Second, the SAML2 library no longer
throws exceptions when more than one attribute value is set on the epti
attribute. This is now logged.

In accordance with the PO, we decided to remove these tests and rely
on EB to never release more than one epti value.

So, in addition to removing the tests, I also dropped the version
restraint on SAML2 3.2

https://github.com/OpenConext/OpenConext-profile/pull/113#issuecomment-465598806