Loosen the assertions on locale configuration
The global view parameters used to verify very strictly if all locale settings where present. Any, and all locales needed to be in the global_view_config.yml.

This is now loosened, as in: All enabled languages need to be translated in the global view config file. But additional items are allowed. Allowing us to just drop in the default g-v-c.yml without error.

This required a change to the custom keysAre assertion. Which was created for this sole purpose. The