Update front controllers for PHP 7
In PHP 7 we no longer have to use the class loader to get a performance
boost. The optimized autoload file is faster from PHP 7 and onward.

For the production front controller, a fallback for pre PHP 7
installations was added in case of a rollback. For the development
version we do not have such a fallback.

The app_dev template was also updated.