Fix deprecation warnings
In order to get rid of much config, autowiring was enabled. Requiring
some additional work to get the controllers to wire correctly.

The extra bundle was only used for the JMS web translation interface
and it was causing deprecation warnings. As we do not use this UI,
removing it seemed like the best solution.