Remove MDL source and move compiled css/js
The MDL package was tracked in the project. Probably in order to simply
upgrade to newer version. A better solution would be to start using NPM
for these kind of dependencies. This is something for the icebox.

By removing the MDL package, the security wanring from Github is
silenced. This was a warning on a dev dependency of the MDL library.