Install Webpack Encore
This is both a Composer and npm package. The composer package integrates
the webpack setup in Symfony. Making it very easy to use the generated
JS/CSS in the application

The NPM dependency does the heavy lifting, and is a wrapper around
Webpack. For now a vanilla webpack config is set, this will be adjusted
in a successive PR when the setup is tailored for OpenConext-Profile