Fix accessibility problems
Prior to this change, accessibility was good, but not thoroughly tested.

This change adds automatic tests using cypress, axe and html-validator & adds those to the CI.
It also fixes all problems found in manual & automatic testing:
- mobile menu didn't work always on tap
- pressing the space bar didn't activate the details element
- something incorrect modal didn't work with the space bar
- some colour contrast issues
- adds focusable to the arrow svg (mandatory attribute)
- add text to the logo-link
- add type to the language-switcher buttons
- add type to the mobile navigation button
- ensure the mobile menu isn't hidden when no js is available
- add a label to the language switcher

To earn a boyscout badge, a problem with the language switcher was also fixed (where it didn't work anymore).