Prevent error log cluttering with Xdebug warnings
The Xdebug option: start with request is warning us that no debug
session was started when the user does not initiate a debug session.
This spams the logs while not debugging.

Removing this option and use the tried and tested debug cookie in the
browser fixes this little inconvenience. Some more details in this
ticket:

https://www.pivotaltracker.com/story/show/179447030