Update to new Twig standards
The twig engine is now used throughout the project, the old
environment solution was discontinued.

Template references have been updated to the new standard using slashes.

The templating reference in the framework config was removed. This
requires an update to the SAML bundle to prevent other warning/issues.

The saml bundle was updated in a successive commit in this branch