Fix problems in IE11
While testing the design for regressions resulting from the upgrade to SF4, it was noticed that the IE11 meta tag was missing.  After adding that tag, the design was tested thoroughly and several extra problems were found.

This change resolves all found issues:
- added the ie meta tag
- used insertBefore instead of prepend to add a label for the select, as prepend is not supported in IE
- added a polyfill for the details element
- rectified styling issues for the details element
- fix styling of download button in IE11
- ensure icons are visible in IE11
- optimize svg code as a result of the above
- ensure surf logo is visible in header and footer for IE11
- optimize width of attributes in table of my services
- fixed view of "something incorrect" for IE11

See https://github.com/javan/details-element-polyfill for the polyfill used