Improve a11y tests & add document language
Prior to this change, we had some rules in html-validate which threw incorrect errors.  As a result we had to them into warnings.
These issues have been resolved by the project maintainer.

This change upgrades html-validate en cypress-html-validate to the new & improved version.  It also ensures the warnings are errors again and adds a specific rule for heading level.
Finally: an error was found with a lack of a document language when running the accessibility tests.  That was resolved.