Fix malfunctioning a11y test
Prior to this change the myservices a11y test failed because we use a link as a button to open/close the modal.  We do this because it allows us to open/close the modal without js, meaning it still works even if JS isn't loaded on the page.

HTML Validator, however, has a rule on by default to prefer native elements over using aria roles.  This is a good rule and up till recently we could only disable it globally.  After filing an issue with cypress-html-validator it is now possible to override the configuration of html-validator on a per-call basis.

This change applies the disabling of the prefer-native-elements rule for the button element on the myservices page.