Add cypress tests to the CI (#201)
* Move cypress to the tests folder

Prior to this change, the cypress tests were in a separate "cypress" folder in the root of the project.  In keeping with the project standard all tests should be in the tests folder.

This change moves the cypres tests in the test folder.

* Add tests to pipeline

* Fix malfunctioning a11y test

Prior to this change the myservices a11y test failed because we use a link as a button to open/close the modal.  We do this because it allows us to open/close the modal without js, meaning it still works even if JS isn't loaded on the page.

HTML Validator, however, has a rule on by default to prefer native elements over using aria roles.  This is a good rule and up till recently we could only disable it globally.  After filing an issue with cypress-html-validator it is now possible to override the configuration of html-validator on a per-call basis.

This change applies the disabling of the prefer-native-elements rule for the button element on the myservices page.

* Docker-compose: Add the profile hostname as alias to the OpenConext image, so other containers can resolve it

* Docker for dev: Move init scripts to /usr/local/sbin
Before they were in /tmp, which is cleaned up after a restart

* Temporarily remove github actions integration

Given that we haven't found a usable work-around for the SAML authentication yet, we are temporarily removing the integration of cypress in the ci.

Co-authored-by: Bart Geesink <bart.geesink@surf.nl>