Ensure givenName is not required for using Profile
Not having the GivenName in the released attribute set would halt
profile in the IntroductionController. Effectively showing an error
right after logging in to Profile.

This chane makes the use of the givenName attrib optional. An additional
change might be required ot make the UX more fluent as the header
element is now only stating 'Hi !' when logging in.

One step at a time!

https://www.pivotaltracker.com/story/show/179882084