Ensure the something incorrect disclosure is shown
Prior to this change, the id for the input element toggling the tooltip was generated incorrectly.  This made all labels of different organisations toggle the same disclosure modal.

This change uses the organizationId as an id for the input.  Thus ensuring that the modal is toggled correctly.

Pivotal ticket: https://www.pivotaltracker.com/story/show/180095824