Create Github Actions release workflow
In this workflow an incomming tag is triggering the makeRelease script.
The output of this script (the artefacts) are used to create a new
release page on the projects github releases page. The artefacts are
then uploaded to that new release.

The one thing remaining is to add the release notes to the release. This
is now set with a small placeholder message