Do not include list of aggregated attributes inside if for idp attributes
The first if checks if the user has any IdP released attributes, and
only then displays those and the aggregated attributes. It is however
possible (though not common) that a service has an empty list of
IdP attributes but does get info from other sources. So move the
aggregation block outside of the if.