No longer run security check on integration tests
There's no relation between changes in a PR and the presence of
a security issue or between the frequency of PRs and the
frequency we want to check for security issues. So remove it
here. We keep the script in bin/ for now, might come in handy
at some point.