Restrict max-width of the logo
Take 2

The max witdh of the contianer combined with the overflow hidden already
caused the image from no longer overlap the text. That was fixed in the
previous commit. But the logo would still be displayed partially. As the
width of the logo was not restricted. The max width is now inherited
from the container <div>.